+++ seed/ts-sdk/exhaustive/inline-path-parameters/src/api/resources/types/resources/object/errors/index.ts
@@ -1,4 +1,0 @@
-export * from "./ObjectWithOptionalFieldError";
-export * from "./ObjectWithRequiredFieldError";
-export * from "./NestedObjectWithOptionalFieldError";
-export * from "./NestedObjectWithRequiredFieldError";